Fix message flag for polls

+++ src/main/java/org/rsa/listeners/MessageListener.java
@@ -71,7 +71,7 @@ public class MessageListener extends ListenerAdapter {
                 String data = rawData.get("d").toString();
                 System.out.println(data);
 
-                if (data.contains("poll=")) {
+                if (data.contains("poll=") && !event.getMessage().getContentRaw().contains("poll=")) {
                     System.out.println("Poll detected, deleting.");
                     event.getMessage().reply("Please do not send polls outside of the poll channel.").queue(m -> {
                         event.getMessage().delete().queue();
